+++ apps/sv/src/main/scala/com/daml/network/sv/onboarding/founder/FoundingNodeInitializer.scala
@@ -370,10 +370,9 @@ class FoundingNodeInitializer(
     private val domainNodeReconciler = new DomainNodeReconciler(
       svcStore,
       svcStoreWithIngestion.connection,
-      config.scan,
-      clock,
-      retryProvider,
-      logger,
+      clock = clock,
+      retryProvider = retryProvider,
+      logger = logger,
     )
 
     /** The one and only entry-point: found a fresh collective, given a properly allocated SVC party */