+++ apps/sv/src/main/scala/com/daml/network/sv/onboarding/joining/JoiningNodeInitializer.scala
@@ -233,7 +233,6 @@ class JoiningNodeInitializer(
     val domainNodeReconciler = new DomainNodeReconciler(
       svcStore,
       svcAutomationService.connection,
-      config.scan,
       clock,
       retryProvider,
       logger,